requirements.txt
fastapi
uvicorn
passlib
python-jose
jinja2
